関数myRootの戻り値がint型のyなので、doubleに修正

+++ main.c
@@ -2,7 +2,7 @@
 #include	<math.h>
 #include	<stdlib.h>
 
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
 {
 	double y;
 	if( x < 0 ){
